requirements.txt
asgiref==3.7.2
astroid==2.15.8
Django==5.0
django-bpmn==0.0.4
django-extensions==3.2.3
django-filter==23.5
exceptiongroup==1.2.0
flake8==6.1.0
flake8-django==1.4
iniconfig==2.0.0
lazy-object-proxy==1.10.0
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pydot==2.0.0
pyflakes==3.1.0
pyparsing==3.1.1
pytest==7.4.4
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
wrapt==1.16.0
